splitting toolbar and figuremanager examples
